SC-8. Add Book, ContactBook, NoteBook classes
