+++ src/handlers/_notes_handler.py
@@ -1,1 +1,5 @@
 # TODO: Add note handler functions. Similar to contacts_handler.py
+# TODO: SC-20. Feature 6. Save Notes
+# TODO: SC-21. Feature 7. Find Notes functionality
+# TODO: SC-22. Feature 8.1 Edit Notes functionality
+# TODO: SC-23. Feature 8.2 Delete Notes functionality
