+++ src/modules/note_main.py
@@ -1,0 +1,6 @@
+from src.modules.field import Field
+
+class Main(Field):
+    def __init__(self, main):
+        super().__init__(main)
+        self.main = Field(main)
